refactor(todo-list): reorganize file structure
