+++ src/todo-list/task-create.command.cli.spec.ts
@@ -3,32 +3,9 @@ import { Logger, MemoryLoggerTransport } from '@deepkit/logger';
 import { Database, MemoryDatabaseAdapter } from '@deepkit/orm';
 import { expect, test } from '@jest/globals';
 
-import { TaskService } from '../app/task.service';
-import { TaskCreateCommand, TaskListCommand } from './task.controller.cli';
-import { Task } from '../app/task.entity';
-
-test('task:list command', async () => {
-  const memoryLogger = new MemoryLoggerTransport();
-  const testing = createTestingApp({
-    controllers: [TaskListCommand],
-    providers: [
-      {
-        provide: Logger,
-        useValue: new Logger([memoryLogger]),
-      },
-      {
-        provide: Database,
-        useFactory: () => new Database(new MemoryDatabaseAdapter(), [Task]),
-      },
-      TaskService,
-    ],
-  });
-
-  await expect(testing.app.execute(['task:list'])).resolves.toBe(0);
-  expect(memoryLogger.messages[0]).toMatchObject({
-    message: '[]',
-  });
-});
+import { TaskCreateCommand } from './task-create.command.cli';
+import { Task } from './task.entity';
+import { TaskService } from './task.service';
 
 test('task:create command', async () => {
   const memoryLogger = new MemoryLoggerTransport();
