+++ src/todo-list/task-create.command.cli.ts
@@ -1,21 +1,8 @@
 import { Command, cli } from '@deepkit/app';
 import { Logger } from '@deepkit/logger';
 
-import { Task } from '../app/task.entity';
-import { TaskService } from '../app/task.service';
-
-@cli.controller('task:list')
-export class TaskListCommand implements Command {
-  constructor(
-    private taskService: TaskService,
-    private logger: Logger,
-  ) {}
-
-  async execute(): Promise<void> {
-    const tasks = await this.taskService.getAll();
-    this.logger.log(tasks);
-  }
-}
+import { Task } from './task.entity';
+import { TaskService } from './task.service';
 
 @cli.controller('task:create')
 export class TaskCreateCommand implements Command {
